friday fun

+++ Util.java
@@ -1,8 +1,8 @@
 public class Util {
     
     public static boolean isPrime(long p) {
-        int sqrt = (int)Math.sqrt(p) + 1;
-        for (int i = 2; i < sqrt; i++) {
+        long sqrt = (long)Math.sqrt(p) + 1;
+        for (long i = 2; i < sqrt; i++) {
             if (p % i == 0)
                 return false;
         }
